+++ src/main/java/de/hdg/keklist/gui/pages/MainPageEvent.java
@@ -1,4 +1,4 @@
-package de.hdg.keklist.gui.events;
+package de.hdg.keklist.gui.pages;
 
 import de.hdg.keklist.Keklist;
 import de.hdg.keklist.gui.GuiManager;
@@ -10,18 +10,18 @@ import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryCloseEvent;
 import org.jetbrains.annotations.NotNull;
 
-public class MainGUIEvent implements Listener {
+public class MainPageEvent implements Listener {
 
     @EventHandler
     public void onClick(@NotNull InventoryClickEvent event) {
         if (event.getCurrentItem() == null) return;
         if (event.getClickedInventory() == null) return;
-        if(!(event.getWhoClicked() instanceof Player player)) return;
+        if (!(event.getWhoClicked() instanceof Player player)) return;
 
-        if(event.getView().title().equals(Keklist.getInstance().getMiniMessage().deserialize("<gold><b>Keklist"))) {
+        if (event.getView().title().equals(Keklist.getInstance().getMiniMessage().deserialize("<gold><b>Keklist"))) {
             event.setCancelled(true);
 
-            if(!player.hasPermission("keklist.gui.use"))
+            if (!player.hasPermission("keklist.gui.use"))
                 player.closeInventory(InventoryCloseEvent.Reason.CANT_USE);
 
             GuiManager.handleMainGUICLick(GuiManager.GuiPage.valueOf(PlainTextComponentSerializer.plainText().serialize(event.getCurrentItem().getItemMeta().displayName()).toUpperCase()), player);
